requirements.txt
streamlit==1.37.0
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
SQLAlchemy==2.0.32
python-dotenv==1.0.1
pydeck==0.9.1
ruamel.yaml==0.17.32
altair==5.5.0
